requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
pydobot==1.3.2
pyserial==3.4
